requirements.txt
flask
firebase-admin
flask-cors
easyocr
numpy
Pillow
torch
transformers
opencv-python
pandas
polars
scikit-learn